requirements.txt
flask==2.0.3
flask-cors==3.0.10 
torch==2.0.1
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1